AngularProject/src/main/webapp/app/entities/portfolio-my-suffix/delete/portfolio-my-suffix-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { PortfolioMySuffixService } from '../service/portfolio-my-suffix.service';

import { PortfolioMySuffixDeleteDialogComponent } from './portfolio-my-suffix-delete-dialog.component';

describe('PortfolioMySuffix Management Delete Component', () => {
  let comp: PortfolioMySuffixDeleteDialogComponent;
  let fixture: ComponentFixture<PortfolioMySuffixDeleteDialogComponent>;
  let service: PortfolioMySuffixService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, PortfolioMySuffixDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(PortfolioMySuffixDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(PortfolioMySuffixDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(PortfolioMySuffixService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
